Podcast_generator_backend/pyproject.toml
[project]
name = "ai-podcast-generator"
version = "0.1.0"
description = "AI-powered podcast generator using GPT-4 and OpenAI TTS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "openai>=1.3.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
